test: added unit tests for Square
