+++ client/package.json
@@ -6,16 +6,11 @@
   "scripts": {
     "start": "NODE_ENV=development react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --coverage",
+    "test": "react-scripts test",
+    "test-coverage": "react-scripts test --coverage",
     "eject": "react-scripts eject",
     "lint": "eslint . --ignore-path .gitignore --ext .js,.jsx,.ts,.tsx --quiet",
     "lint-fix": "eslint .  --ignore-path .gitignore --ext .js,.jsx,.ts,.tsx --fix"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
   },
   "dependencies": {
     "@types/uuid": "^8.3.0",
@@ -40,14 +35,28 @@
     "@types/react": "^17.0.2",
     "@types/react-dom": "^17.0.1",
     "@types/react-router-dom": "^5.1.7",
+    "@types/react-test-renderer": "^17.0.1",
     "@typescript-eslint/eslint-plugin": "^4.13.0",
     "@typescript-eslint/parser": "^4.13.0",
     "eslint": "^7.17.0",
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-simple-import-sort": "^7.0.0",
+    "identity-obj-proxy": "^3.0.0",
     "prettier": "^2.2.1",
+    "react-test-renderer": "^17.0.1",
     "typescript": "^4.1.3"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(css|less)$": "identity-obj-proxy"
+    }
   },
   "browserslist": {
     "production": [
