+++ client/src/tests/context/stores/mock/boardWithOneMoveMock.ts
@@ -1,15 +1,8 @@
-import { TileValue } from '../../../types/game';
-// this not a finished board
-/*
- with the following situation: ( . = not finished )
- X X .
- . O .
- . O .
- */
+import { TileValue } from '../../../../types/game';
 
-const unfinishedBoardMock = [
+export const boardWithOneMoveMock = [
   {
-    value: TileValue.Cross,
+    value: TileValue.Empty,
     position: {
       x: 0,
       y: 0
@@ -46,7 +39,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 2
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -57,7 +50,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 0
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -68,7 +61,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -79,7 +72,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -117,7 +110,7 @@ const unfinishedBoardMock = [
     ]
   },
   {
-    value: TileValue.Cross,
+    value: TileValue.Empty,
     position: {
       x: 0,
       y: 1
@@ -165,7 +158,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 0
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -176,7 +169,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -187,7 +180,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -262,7 +255,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 2
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -284,7 +277,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -348,7 +341,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 0
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -359,7 +352,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 1
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -392,7 +385,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -403,7 +396,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -425,7 +418,7 @@ const unfinishedBoardMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -441,7 +434,7 @@ const unfinishedBoardMock = [
     ]
   },
   {
-    value: TileValue.Circle,
+    value: TileValue.Empty,
     position: {
       x: 1,
       y: 1
@@ -456,7 +449,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 0
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -467,7 +460,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 1
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -478,7 +471,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 2
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -489,7 +482,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 0
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -522,7 +515,7 @@ const unfinishedBoardMock = [
           x: 2,
           y: 0
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -564,7 +557,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 0
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -575,7 +568,7 @@ const unfinishedBoardMock = [
           x: 0,
           y: 1
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -608,7 +601,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -641,7 +634,7 @@ const unfinishedBoardMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -749,7 +742,7 @@ const unfinishedBoardMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -765,7 +758,7 @@ const unfinishedBoardMock = [
     ]
   },
   {
-    value: TileValue.Circle,
+    value: TileValue.Empty,
     position: {
       x: 2,
       y: 1
@@ -813,7 +806,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 0
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -824,7 +817,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -835,7 +828,7 @@ const unfinishedBoardMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Circle
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -965,7 +958,7 @@ const unfinishedBoardMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Cross
+        value: TileValue.Empty
       },
       {
         boardPosition: {
@@ -982,331 +975,9 @@ const unfinishedBoardMock = [
   }
 ];
 
-const movesForUnfinishedBoardMock = [
+export const activeBoardsForBoardWithOneMoveMock = [
   {
-    boardPosition: {
-      x: 0,
-      y: 0
-    },
-    tilePosition: {
-      x: 1,
-      y: 1
-    },
-    player: 0,
-    moveNumber: 1
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 1
-    },
-    tilePosition: {
-      x: 0,
-      y: 0
-    },
-    player: 1,
-    moveNumber: 2
-  },
-  {
-    boardPosition: {
-      x: 0,
-      y: 0
-    },
-    tilePosition: {
-      x: 1,
-      y: 2
-    },
-    player: 0,
-    moveNumber: 3
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 2
-    },
-    tilePosition: {
-      x: 0,
-      y: 0
-    },
-    player: 1,
-    moveNumber: 4
-  },
-  {
-    boardPosition: {
-      x: 0,
-      y: 0
-    },
-    tilePosition: {
-      x: 1,
-      y: 0
-    },
-    player: 0,
-    moveNumber: 5
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 0
-    },
-    tilePosition: {
-      x: 0,
-      y: 1
-    },
-    player: 1,
-    moveNumber: 6
-  },
-  {
-    boardPosition: {
-      x: 0,
-      y: 1
-    },
-    tilePosition: {
-      x: 1,
-      y: 1
-    },
-    player: 0,
-    moveNumber: 7
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 1
-    },
-    tilePosition: {
-      x: 0,
-      y: 1
-    },
-    player: 1,
-    moveNumber: 8
-  },
-  {
-    boardPosition: {
-      x: 0,
-      y: 1
-    },
-    tilePosition: {
-      x: 1,
-      y: 2
-    },
-    player: 0,
-    moveNumber: 9
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 2
-    },
-    tilePosition: {
-      x: 0,
-      y: 1
-    },
-    player: 1,
-    moveNumber: 10
-  },
-  {
-    boardPosition: {
-      x: 0,
-      y: 1
-    },
-    tilePosition: {
-      x: 1,
-      y: 0
-    },
-    player: 0,
-    moveNumber: 11
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 0
-    },
-    tilePosition: {
-      x: 1,
-      y: 1
-    },
-    player: 1,
-    moveNumber: 12
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 1
-    },
-    tilePosition: {
-      x: 0,
-      y: 2
-    },
-    player: 0,
-    moveNumber: 13
-  },
-  {
-    boardPosition: {
-      x: 0,
-      y: 2
-    },
-    tilePosition: {
-      x: 0,
-      y: 2
-    },
-    player: 1,
-    moveNumber: 14
-  },
-  {
-    boardPosition: {
-      x: 0,
-      y: 2
-    },
-    tilePosition: {
-      x: 1,
-      y: 1
-    },
-    player: 0,
-    moveNumber: 15
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 1
-    },
-    tilePosition: {
-      x: 1,
-      y: 0
-    },
-    player: 1,
-    moveNumber: 16
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 0
-    },
-    tilePosition: {
-      x: 0,
-      y: 0
-    },
-    player: 0,
-    moveNumber: 17
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 1
-    },
-    tilePosition: {
-      x: 2,
-      y: 0
-    },
-    player: 1,
-    moveNumber: 18
-  },
-  {
-    boardPosition: {
-      x: 2,
-      y: 0
-    },
-    tilePosition: {
-      x: 2,
-      y: 1
-    },
-    player: 0,
-    moveNumber: 19
-  },
-  {
-    boardPosition: {
-      x: 2,
-      y: 1
-    },
-    tilePosition: {
-      x: 1,
-      y: 1
-    },
-    player: 1,
-    moveNumber: 20
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 0
-    },
-    tilePosition: {
-      x: 2,
-      y: 1
-    },
-    player: 0,
-    moveNumber: 21
-  },
-  {
-    boardPosition: {
-      x: 2,
-      y: 1
-    },
-    tilePosition: {
-      x: 1,
-      y: 0
-    },
-    player: 1,
-    moveNumber: 22
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 0
-    },
-    tilePosition: {
-      x: 1,
-      y: 2
-    },
-    player: 0,
-    moveNumber: 23
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 2
-    },
-    tilePosition: {
-      x: 1,
-      y: 1
-    },
-    player: 1,
-    moveNumber: 24
-  },
-  {
-    boardPosition: {
-      x: 2,
-      y: 2
-    },
-    tilePosition: {
-      x: 2,
-      y: 1
-    },
-    player: 0,
-    moveNumber: 25
-  },
-  {
-    boardPosition: {
-      x: 2,
-      y: 1
-    },
-    tilePosition: {
-      x: 1,
-      y: 2
-    },
-    player: 1,
-    moveNumber: 26
-  },
-  {
-    boardPosition: {
-      x: 1,
-      y: 2
-    },
-    tilePosition: {
-      x: 2,
-      y: 1
-    },
-    player: 0,
-    moveNumber: 27
+    x: 0,
+    y: 2
   }
 ];
-
-export { movesForUnfinishedBoardMock, unfinishedBoardMock };
