+++ client/src/tests/context/stores/mock/unfinishedBoardMock.ts
@@ -1,8 +1,15 @@
-import { TileValue } from '../../../types/game';
+import { TileValue } from '../../../../types/game';
+// this not a finished board
+/*
+ with the following situation: ( . = not finished )
+ X X .
+ . O .
+ . O .
+ */
 
-export const boardWithTwoMovesMock = [
+const unfinishedBoardMock = [
   {
-    value: TileValue.Empty,
+    value: TileValue.Cross,
     position: {
       x: 0,
       y: 0
@@ -38,6 +45,17 @@ export const boardWithTwoMovesMock = [
         position: {
           x: 0,
           y: 2
+        },
+        value: TileValue.Empty
+      },
+      {
+        boardPosition: {
+          x: 0,
+          y: 0
+        },
+        position: {
+          x: 1,
+          y: 0
         },
         value: TileValue.Cross
       },
@@ -48,20 +66,9 @@ export const boardWithTwoMovesMock = [
         },
         position: {
           x: 1,
-          y: 0
-        },
-        value: TileValue.Empty
-      },
-      {
-        boardPosition: {
-          x: 0,
-          y: 0
-        },
-        position: {
-          x: 1,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -72,7 +79,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -110,7 +117,7 @@ export const boardWithTwoMovesMock = [
     ]
   },
   {
-    value: TileValue.Empty,
+    value: TileValue.Cross,
     position: {
       x: 0,
       y: 1
@@ -158,7 +165,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 0
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -169,7 +176,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -180,7 +187,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -254,17 +261,6 @@ export const boardWithTwoMovesMock = [
         position: {
           x: 0,
           y: 2
-        },
-        value: TileValue.Empty
-      },
-      {
-        boardPosition: {
-          x: 0,
-          y: 2
-        },
-        position: {
-          x: 1,
-          y: 0
         },
         value: TileValue.Circle
       },
@@ -275,9 +271,20 @@ export const boardWithTwoMovesMock = [
         },
         position: {
           x: 1,
-          y: 1
+          y: 0
         },
         value: TileValue.Empty
+      },
+      {
+        boardPosition: {
+          x: 0,
+          y: 2
+        },
+        position: {
+          x: 1,
+          y: 1
+        },
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -341,7 +348,7 @@ export const boardWithTwoMovesMock = [
           x: 0,
           y: 0
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -352,7 +359,7 @@ export const boardWithTwoMovesMock = [
           x: 0,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -385,7 +392,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -396,7 +403,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -418,7 +425,7 @@ export const boardWithTwoMovesMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -434,7 +441,7 @@ export const boardWithTwoMovesMock = [
     ]
   },
   {
-    value: TileValue.Empty,
+    value: TileValue.Circle,
     position: {
       x: 1,
       y: 1
@@ -449,7 +456,7 @@ export const boardWithTwoMovesMock = [
           x: 0,
           y: 0
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -460,7 +467,7 @@ export const boardWithTwoMovesMock = [
           x: 0,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -471,7 +478,7 @@ export const boardWithTwoMovesMock = [
           x: 0,
           y: 2
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -482,7 +489,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 0
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -515,7 +522,7 @@ export const boardWithTwoMovesMock = [
           x: 2,
           y: 0
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -557,7 +564,7 @@ export const boardWithTwoMovesMock = [
           x: 0,
           y: 0
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -568,7 +575,7 @@ export const boardWithTwoMovesMock = [
           x: 0,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -601,7 +608,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -634,7 +641,7 @@ export const boardWithTwoMovesMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -742,7 +749,7 @@ export const boardWithTwoMovesMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -758,7 +765,7 @@ export const boardWithTwoMovesMock = [
     ]
   },
   {
-    value: TileValue.Empty,
+    value: TileValue.Circle,
     position: {
       x: 2,
       y: 1
@@ -806,7 +813,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 0
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -817,7 +824,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -828,7 +835,7 @@ export const boardWithTwoMovesMock = [
           x: 1,
           y: 2
         },
-        value: TileValue.Empty
+        value: TileValue.Circle
       },
       {
         boardPosition: {
@@ -958,7 +965,7 @@ export const boardWithTwoMovesMock = [
           x: 2,
           y: 1
         },
-        value: TileValue.Empty
+        value: TileValue.Cross
       },
       {
         boardPosition: {
@@ -975,9 +982,331 @@ export const boardWithTwoMovesMock = [
   }
 ];
 
-export const activeBoardsForBoardWithTwoMovesMock = [
+const movesForUnfinishedBoardMock = [
   {
-    x: 1,
-    y: 0
+    boardPosition: {
+      x: 0,
+      y: 0
+    },
+    tilePosition: {
+      x: 1,
+      y: 1
+    },
+    player: 0,
+    moveNumber: 1
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 1
+    },
+    tilePosition: {
+      x: 0,
+      y: 0
+    },
+    player: 1,
+    moveNumber: 2
+  },
+  {
+    boardPosition: {
+      x: 0,
+      y: 0
+    },
+    tilePosition: {
+      x: 1,
+      y: 2
+    },
+    player: 0,
+    moveNumber: 3
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 2
+    },
+    tilePosition: {
+      x: 0,
+      y: 0
+    },
+    player: 1,
+    moveNumber: 4
+  },
+  {
+    boardPosition: {
+      x: 0,
+      y: 0
+    },
+    tilePosition: {
+      x: 1,
+      y: 0
+    },
+    player: 0,
+    moveNumber: 5
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 0
+    },
+    tilePosition: {
+      x: 0,
+      y: 1
+    },
+    player: 1,
+    moveNumber: 6
+  },
+  {
+    boardPosition: {
+      x: 0,
+      y: 1
+    },
+    tilePosition: {
+      x: 1,
+      y: 1
+    },
+    player: 0,
+    moveNumber: 7
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 1
+    },
+    tilePosition: {
+      x: 0,
+      y: 1
+    },
+    player: 1,
+    moveNumber: 8
+  },
+  {
+    boardPosition: {
+      x: 0,
+      y: 1
+    },
+    tilePosition: {
+      x: 1,
+      y: 2
+    },
+    player: 0,
+    moveNumber: 9
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 2
+    },
+    tilePosition: {
+      x: 0,
+      y: 1
+    },
+    player: 1,
+    moveNumber: 10
+  },
+  {
+    boardPosition: {
+      x: 0,
+      y: 1
+    },
+    tilePosition: {
+      x: 1,
+      y: 0
+    },
+    player: 0,
+    moveNumber: 11
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 0
+    },
+    tilePosition: {
+      x: 1,
+      y: 1
+    },
+    player: 1,
+    moveNumber: 12
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 1
+    },
+    tilePosition: {
+      x: 0,
+      y: 2
+    },
+    player: 0,
+    moveNumber: 13
+  },
+  {
+    boardPosition: {
+      x: 0,
+      y: 2
+    },
+    tilePosition: {
+      x: 0,
+      y: 2
+    },
+    player: 1,
+    moveNumber: 14
+  },
+  {
+    boardPosition: {
+      x: 0,
+      y: 2
+    },
+    tilePosition: {
+      x: 1,
+      y: 1
+    },
+    player: 0,
+    moveNumber: 15
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 1
+    },
+    tilePosition: {
+      x: 1,
+      y: 0
+    },
+    player: 1,
+    moveNumber: 16
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 0
+    },
+    tilePosition: {
+      x: 0,
+      y: 0
+    },
+    player: 0,
+    moveNumber: 17
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 1
+    },
+    tilePosition: {
+      x: 2,
+      y: 0
+    },
+    player: 1,
+    moveNumber: 18
+  },
+  {
+    boardPosition: {
+      x: 2,
+      y: 0
+    },
+    tilePosition: {
+      x: 2,
+      y: 1
+    },
+    player: 0,
+    moveNumber: 19
+  },
+  {
+    boardPosition: {
+      x: 2,
+      y: 1
+    },
+    tilePosition: {
+      x: 1,
+      y: 1
+    },
+    player: 1,
+    moveNumber: 20
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 0
+    },
+    tilePosition: {
+      x: 2,
+      y: 1
+    },
+    player: 0,
+    moveNumber: 21
+  },
+  {
+    boardPosition: {
+      x: 2,
+      y: 1
+    },
+    tilePosition: {
+      x: 1,
+      y: 0
+    },
+    player: 1,
+    moveNumber: 22
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 0
+    },
+    tilePosition: {
+      x: 1,
+      y: 2
+    },
+    player: 0,
+    moveNumber: 23
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 2
+    },
+    tilePosition: {
+      x: 1,
+      y: 1
+    },
+    player: 1,
+    moveNumber: 24
+  },
+  {
+    boardPosition: {
+      x: 2,
+      y: 2
+    },
+    tilePosition: {
+      x: 2,
+      y: 1
+    },
+    player: 0,
+    moveNumber: 25
+  },
+  {
+    boardPosition: {
+      x: 2,
+      y: 1
+    },
+    tilePosition: {
+      x: 1,
+      y: 2
+    },
+    player: 1,
+    moveNumber: 26
+  },
+  {
+    boardPosition: {
+      x: 1,
+      y: 2
+    },
+    tilePosition: {
+      x: 2,
+      y: 1
+    },
+    player: 0,
+    moveNumber: 27
   }
 ];
+
+export { movesForUnfinishedBoardMock, unfinishedBoardMock };
