+++ client/tsconfig.json
@@ -19,5 +19,5 @@
     "jsx": "react-jsx",
     "types": ["jest", "node"]
   },
-  "include": ["src"]
+  "include": ["src", "src/setupTests.ts"]
 }
